+++ NEWS.md
@@ -1,3 +1,5 @@
+# websocket (development version)
+
 # websocket 1.4.1
 
 * Add UCRT toolchain support (@jeroen, #82)